+++ Strata/Source/LevelExcavationUtility.cs
@@ -30,7 +30,9 @@ namespace Strata
             }
             foreach (Thing thing in map.listerThings.ThingsInGroup(ThingRequestGroup.MapPortal))
             {
-                if (thing == ignore || thing is not Building_StairsDown down || thing is Building_AncientColonyStairsDown
+                if (thing == ignore || thing is not Building_StairsDown down
+                    || thing is Building_StairsBuildUp
+                    || thing is Building_AncientColonyStairsDown
                     || !down.Spawned || !down.PocketMapExists)
                 {
                     continue;